* PHP-8.3:
  Fix GH-20732: Phar::LoadPhar undefined behavior when loading directory

+++ ext/phar/phar.c
@@ -1611,7 +1611,7 @@ static zend_result phar_open_from_fp(php_stream* fp, char *fname, size_t fname_l
 	const zend_long readsize = sizeof(buffer) - sizeof(token);
 	const zend_long tokenlen = sizeof(token) - 1;
 	zend_long halt_offset;
-	size_t got;
+	ssize_t got;
 	uint32_t compression = PHAR_FILE_COMPRESSED_NONE;
 
 	if (error) {
@@ -1629,7 +1629,7 @@ static zend_result phar_open_from_fp(php_stream* fp, char *fname, size_t fname_l
 	/* Maybe it's better to compile the file instead of just searching,  */
 	/* but we only want the offset. So we want a .re scanner to find it. */
 	while(!php_stream_eof(fp)) {
-		if ((got = php_stream_read(fp, buffer+tokenlen, readsize)) < (size_t) tokenlen) {
+		if ((got = php_stream_read(fp, buffer+tokenlen, readsize)) < tokenlen) {
 			MAPPHAR_ALLOC_FAIL("internal corruption of phar \"%s\" (truncated entry)")
 		}
